c++ 문법에 맞게 struct 구초제 수정 Getter함수가 float 변수를 int로 반환하는 것을 수정 경고 메세지로 인해 변수 초기화 부분 추가.

+++ Find To Hensel/Includes/FindToHensel/Object/Enemy/Enemy.h
@@ -9,8 +9,8 @@ private:
 	float speed = 0.f;
 	int objectcode = 0;
 	int imagecode = 0;
-	D3DXVECTOR2 pos = D3DXVECTOR2(0.f,0.f);
-	int score=0;
+	D3DXVECTOR2 pos = D3DXVECTOR2(0.f, 0.f);
+	int score = 0;
 
 protected:
 	float time = 0.f;
@@ -36,7 +36,7 @@ public:
 	virtual void damaged(int);
 	const int& getImg() const;
 	const int& getScore( ) const;
-	const int& getSpeed() const;
+	const float& getSpeed() const;
 	const float& getX() const;
 	const float& getY() const;
 	const D3DXVECTOR2& getPos() const;
